+++ src/utils/ask.ts
@@ -3,7 +3,7 @@ import { showLogsCommand, logMessage } from "../utils/log";
 
 export async function askTrustAuthors(): Promise<number> {
 	const config = vscode.workspace.getConfiguration("quartoWizard.ask", null);
-	let configTrustAuthors = config.get<string>("trustAuthors");
+	const configTrustAuthors = config.get<string>("trustAuthors");
 
 	if (configTrustAuthors === "ask") {
 		const trustAuthors = await vscode.window.showQuickPick(
@@ -21,7 +21,7 @@ export async function askTrustAuthors(): Promise<number> {
 			return 0;
 		} else if (trustAuthors?.label !== "Yes") {
 			const message = "Operation cancelled because the authors are not trusted.";
-			logMessage(message);
+			logMessage(message, "info");
 			vscode.window.showInformationMessage(`${message} ${showLogsCommand()}.`);
 			return 1;
 		}
@@ -31,7 +31,7 @@ export async function askTrustAuthors(): Promise<number> {
 
 export async function askConfirmInstall(): Promise<number> {
 	const config = vscode.workspace.getConfiguration("quartoWizard.ask", null);
-	let configConfirmInstall = config.get<string>("confirmInstall");
+	const configConfirmInstall = config.get<string>("confirmInstall");
 
 	if (configConfirmInstall === "ask") {
 		const installWorkspace = await vscode.window.showQuickPick(
@@ -49,7 +49,7 @@ export async function askConfirmInstall(): Promise<number> {
 			return 0;
 		} else if (installWorkspace?.label !== "Yes") {
 			const message = "Operation cancelled by the user.";
-			logMessage(message);
+			logMessage(message, "info");
 			vscode.window.showInformationMessage(`${message} ${showLogsCommand()}.`);
 			return 1;
 		}
@@ -59,7 +59,7 @@ export async function askConfirmInstall(): Promise<number> {
 
 export async function askConfirmRemove(): Promise<number> {
 	const config = vscode.workspace.getConfiguration("quartoWizard.ask");
-	let configConfirmRemove = config.get<string>("confirmRemove");
+	const configConfirmRemove = config.get<string>("confirmRemove");
 
 	if (configConfirmRemove === "always") {
 		const removeWorkspace = await vscode.window.showQuickPick(
@@ -77,7 +77,7 @@ export async function askConfirmRemove(): Promise<number> {
 			return 0;
 		} else if (removeWorkspace?.label !== "Yes") {
 			const message = "Operation cancelled by the user.";
-			logMessage(message);
+			logMessage(message, "info");
 			vscode.window.showInformationMessage(`${message} ${showLogsCommand()}.`);
 			return 1;
 		}